# doc/swagger.yaml
openapi: 3.0.0
paths:
...
  /article:
    post:
      operationId: ArticleController_create
      summary: Create article
      description: Creates a new article
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateArticleDto'
      responses:
        '201':
          description: The article has been created
          content:
            application/json:
              example:
                id: 0e2f864e-8611-4e41-9d49-1c94b1df5c95
                title: string
                content: string
                status: draft
                authorId: null
                categoryId: null
                tags:
                  - nestjs
                  - typescript
                  - backend
                createdAt: 1775299578964
                updatedAt: 1775299578964
        '400':
          description: Bad request. Body does not contain required fields
        '403':
          description: Insufficient permissions
      tags:
        - Articles
    get:
      operationId: ArticleController_findAll
      summary: Get all articles
      description: >-
        Gets all articles. Supports filtering by status, categoryId, and tag,
        pagination and sorting
      parameters:
        - name: page
          required: false
          in: query
          description: Page number (starting from 1)
          schema:
            example: 1
            type: number
        - name: limit
          required: false
          in: query
          description: Number of items per page
          schema:
            example: 10
            type: number
        - name: status
          required: false
          in: query
          description: Filter by article status
          schema:
            example: draft
            enum:
              - archived
              - draft
              - published
            type: string
        - name: categoryId
          required: false
          in: query
          description: Filter by categoryId
          schema:
            example: f3d2f4c6-5376-48df-b4a7-9fe825559db9
            type: string
        - name: tag
          required: false
          in: query
          description: Filter by tag
          schema:
            example: string
            type: string
        - name: sortBy
          required: false
          in: query
          description: The field to sort by
          schema:
            example: title
            enum:
              - id
              - title
              - content
              - status
              - authorId
              - categoryId
              - createdAt
              - updatedAt
            type: string
        - name: order
          required: false
          in: query
          description: Sorting direction
          schema:
            example: DESC
            enum:
              - ASC
              - DESC
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              examples:
                Without pagination:
                  value:
                    - id: 0e2f864e-8611-4e41-9d49-1c94b1df5c95
                      title: string
                      content: string
                      status: draft
                      authorId: null
                      categoryId: null
                      tags:
                        - nestjs
                        - typescript
                        - backend
                      createdAt: 1775299578964
                      updatedAt: 1775299578964
                  description: Without page and limit query parameters
                With pagination:
                  value:
                    total: 1
                    page: 1
                    limit: 10
                    data:
                      - id: 0e2f864e-8611-4e41-9d49-1c94b1df5c95
                        title: string
                        content: string
                        status: draft
                        authorId: null
                        categoryId: null
                        tags:
                          - nestjs
                          - typescript
                          - backend
                        createdAt: 1775299578964
                        updatedAt: 1775299578964
                  description: With page and limit query parameters
        '400':
          description: Bad request. Wrong query parameters (hacker scope)
        '403':
          description: Insufficient permissions
      tags:
        - Articles
  /article/{id}:
    get:
      operationId: ArticleController_findOne
      summary: Get single article by id
      description: Get single article by id
      parameters:
        - name: id
          required: true
          in: path
          description: Article id, format UUID v4
          schema: {}
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              example:
                id: 0e2f864e-8611-4e41-9d49-1c94b1df5c95
                title: string
                content: string
                status: draft
                authorId: null
                categoryId: null
                tags:
                  - nestjs
                  - typescript
                  - backend
                createdAt: 1775299578964
                updatedAt: 1775299578964
        '400':
          description: Bad request. Article id is invalid (not uuid)
        '403':
          description: Insufficient permissions
        '404':
          description: Article not found
      tags:
        - Articles
    put:
      operationId: ArticleController_update
      summary: Update article information
      description: Updates article by ID
      parameters:
        - name: id
          required: true
          in: path
          description: Article id, format UUID v4
          schema: {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateArticleDto'
      responses:
        '200':
          description: The article has been updated
          content:
            application/json:
              example:
                id: 0e2f864e-8611-4e41-9d49-1c94b1df5c95
                title: string
                content: string
                status: draft
                authorId: null
                categoryId: null
                tags:
                  - nestjs
                  - typescript
                  - backend
                createdAt: 1775299578964
                updatedAt: 1775300238971
        '400':
          description: Bad request. Article id is invalid (not uuid)
        '403':
          description: Insufficient permissions
        '404':
          description: Article not found
      tags:
        - Articles
    delete:
      operationId: ArticleController_remove
      summary: Deletes article
      description: Deletes article. Deletes all associated comments.
      parameters:
        - name: id
          required: true
          in: path
          description: Article id, format UUID v4
          schema: {}
      responses:
        '204':
          description: The article has been deleted
          example: ''
        '400':
          description: Bad request. Article id is invalid (not uuid)
        '403':
          description: Insufficient permissions
        '404':
          description: Article not found
      tags:
        - Articles
...
info:
  title: Knowledge Hub
  description: Knowledge hub service for managing articles, categories, and comments
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    CreateArticleDto:
      type: object
      properties:
        title:
          type: string
          example: string
        content:
          type: string
          example: string
        status:
          type: string
          description: Article status
          enum:
            - archived
            - draft
            - published
          example: draft
        authorId:
          type: string
          example: null
        categoryId:
          type: string
          example: null
        tags:
          description: 'Article tags list '
          example:
            - nestjs
            - typescript
            - backend
          type: array
          items:
            type: string
      required:
        - title
        - content
    UpdateArticleDto:
      type: object
      properties:
        title:
          type: string
          example: string
        content:
          type: string
          example: string
        status:
          type: string
          description: Article status
          enum:
            - archived
            - draft
            - published
          example: draft
        authorId:
          type: string
          example: null
        categoryId:
          type: string
          example: null
        tags:
          description: 'Article tags list '
          example:
            - nestjs
            - typescript
            - backend
          type: array
          items:
            type: string
...